chore: create tea bag
