+++ app/api/tea-bag/nft/route.ts
@@ -1,8 +1,8 @@
-import {auth, getSession} from "@/auth";
-import {NextAuthRequest} from "next-auth/lib";
-import {notAuthenticatedProblem, problem} from "@/http/problem";
-import {nftService} from "@/services";
-import {NextResponse} from "next/server";
+import {auth, getSession} from "@/auth"
+import {NextAuthRequest} from "next-auth/lib"
+import {notAuthenticatedProblem, problem} from "@/http/problem"
+import {nftService} from "@/services"
+import {NextResponse} from "next/server"
 
 export const GET = auth(async (req: NextAuthRequest) => {
   const session = getSession(req)
