+++ app/api/tea-bag/report/route.ts
@@ -1,4 +1,5 @@
 import {auth, getSession} from "@/auth"
+// @ts-expect-error TODO fix library not found
 import {NextAuthRequest} from "next-auth/lib"
 import {invalidQueryParameterProblem, notAuthenticatedProblem, problem} from "@/http/problem"
 import {reportProfileService} from "@/services"
