+++ next.config.mjs
@@ -1,6 +1,20 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+/*  reactStrictMode: true,
+  webpack: (webpackConfig, { webpack }) => {
+    webpackConfig.plugins.push(
+      // Remove node: from import specifiers, because Next.js does not yet support node: scheme
+      // https://github.com/vercel/next.js/issues/28774
+      new webpack.NormalModuleReplacementPlugin(
+        /^node:/,
+        (resource) => {
+          resource.request = resource.request.replace(/^node:/, '');
+        },
+      ),
+    );
 
+    return webpackConfig;
+  },*/
 };
 
 export default nextConfig;
