+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "NODE_ENV=production next build",
+    "build": "next build",
     "start": "NODE_ENV=production next start",
     "lint": "next lint",
     "migrate": "knex migrate:latest --esm --cwd ./db_migrations",
@@ -98,6 +98,7 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jest": "^27.9.0",
     "eslint-plugin-prettier": "^5.1.3",
+    "node-polyfill-webpack-plugin": "^4.0.0",
     "pg": "^8.11.5",
     "postcss": "^8.4.38",
     "prettier": "^3.2.5",
