+++ src/services/index.ts
@@ -16,8 +16,11 @@ import {
   profilePageSize,
   searchFollowersProfileSize,
   searchFollowsProfileSize,
-  searchRequesterProfileSize
+  searchRequesterProfileSize,
+  teaBagsPageSize
 } from "@/services/constants"
+import {DefaultTeaBagService} from "@/teaBag/service"
+import {DefaultReportProfileService} from "@/Report-profile/service"
 
 export const userService = new DefaultUserService(pgClient, env.PEPPER_PASSWORD_SECRET, env.TOTP_ENCRYPTION_KEY)
 
@@ -29,13 +32,17 @@ export const profileService = new DefaultProfileService(pgClient, s3client, env.
 
 export const nftService = new DefaultNftService(pgClient, nftsPageSize)
 
+export const teaBagService = new DefaultTeaBagService(pgClient, s3client, env.S3_BUCKET_NAME, teaBagsPageSize)
+
+export const reportProfileService = new DefaultReportProfileService(pgClient)
+
 export const followService = new DefaultFollowService(
-    pgClient,
-    followersPageSize,
-    followsPageSize,
-    followRequestPageSize,
-    followRequestIgnoredPageSize,
-    searchRequesterProfileSize,
-    searchFollowsProfileSize,
-    searchFollowersProfileSize
+  pgClient,
+  followersPageSize,
+  followsPageSize,
+  followRequestPageSize,
+  followRequestIgnoredPageSize,
+  searchRequesterProfileSize,
+  searchFollowsProfileSize,
+  searchFollowersProfileSize
 )
